outputs/pyproject.toml
[project]
name = "autopipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [ "black>=25.1.0", "fastapi>=0.115.12", "isort>=6.0.1", "jupyter>=1.1.1", "matplotlib>=3.10.1", "numpy>=2.2.5", "pandas>=2.2.3", "pillow>=0", "pydantic>=2.11.5", "pyside6>=6.9.0", "qt-material>=2.17", "scikit-learn>=1.6.1", "thefuzz[speedup]>=0.22.1", "toml>=0.10.2", "tool>=0.8.0", "torch>=2.6.0", "torchaudio>=2.6.0", "torchmetrics>=1.7.4", "torchvision>=0.21.0", "uvicorn>=0.34.2",]

[tool.uv]
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
[[tool.uv.sources.torch]]
index = "pytorch-cpu"
marker = "sys_platform != 'linux'"

[[tool.uv.sources.torch]]
index = "pytorch-cu124"
marker = "sys_platform == 'linux'"

[[tool.uv.sources.torchvision]]
index = "pytorch-cpu"
marker = "sys_platform != 'linux'"

[[tool.uv.sources.torchvision]]
index = "pytorch-cu124"
marker = "sys_platform == 'linux'"

[[tool.uv.sources.torchaudio]]
index = "pytorch-cpu"
marker = "sys_platform != 'linux'"

[[tool.uv.sources.torchaudio]]
index = "pytorch-cu124"
marker = "sys_platform == 'linux'"

